Type change

+++ 0x1A-hash_tables/0-hash_table_create.c
@@ -9,7 +9,7 @@
  */
 hash_table_t *hash_table_create(unsigned long int size)
 {
-	int a;
+	unsigned long int a;
 	hash_table_t *hash_t;
 
 	hash_t = malloc(sizeof(hash_table_t));
